add thunk creators to actions folder
