+++ react-burger/src/components/burger-constructor/burger-constructor.jsx
@@ -34,7 +34,7 @@ const BurgerConstructor = ({ handleRequest }) => {
                 <ConstructorElement
                   type="top"
                   isLocked={true}
-                  text={bun.name}
+                  text={`${bun.name} (верх)`}
                   price={bun.price}
                   thumbnail={bun.image}
                 />
@@ -59,7 +59,7 @@ const BurgerConstructor = ({ handleRequest }) => {
                 <ConstructorElement
                   type="bottom"
                   isLocked={true}
-                  text={bun.name}
+                  text={`${bun.name} (низ)`}
                   price={bun.price}
                   thumbnail={bun.image}
                 />
