+++ react-burger/src/services/actions/constructor-actions.js
@@ -1,3 +1,4 @@
+import { BASE_URL } from "../../utils/constants"
 import { ADD_BUN, ADD_INGREDIENT, MOVE_INGREDIENT, REMOVE_ITEM, ADD_TO_INGREDIENTS_MAP, ADD_TO_BUN_MAP, POST_ORDER_SUCCESS, POST_ORDER_REQUEST, POST_ORDER_ERROR } from "../../utils/constants"
 
 export const addBun = item => ({ type: ADD_BUN, item })
@@ -11,3 +12,36 @@ export const removeItem = id => ({ type: REMOVE_ITEM, id })
 export const postOrderSuccess = number => ({ type: POST_ORDER_SUCCESS, number })
 export const postOrderRequest = () => ({ type: POST_ORDER_REQUEST })
 export const postOrderError = () => ({ type: POST_ORDER_ERROR })
+
+export const addToConstructor = item => dispatch => {
+  if (item.type === 'bun') {
+    dispatch(addToBunMap(item._id))
+    dispatch(addBun(item))
+  } else {
+    dispatch(addToIngredientsMap(item._id))
+    dispatch(addIngredient(item))
+  }
+}
+
+export const postOrder = data => dispatch => {
+  dispatch(postOrderRequest())
+  fetch(`${BASE_URL}/orders`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({
+      'ingredients': data
+    }),
+  })
+    .then(res => res.json())
+    .then(res => {
+      if (res && res.success) {
+        dispatch(postOrderSuccess(res.order.number));
+      }
+    })
+    .catch(err => {
+      dispatch(postOrderError())
+      console.log(err)
+    })
+}
