+++ react-burger/src/services/redusers/constructor-reducer.js
@@ -1,4 +1,4 @@
-import { BASE_URL, ADD_BUN, ADD_INGREDIENT, MOVE_INGREDIENT, REMOVE_ITEM, ADD_TO_INGREDIENTS_MAP, ADD_TO_BUN_MAP, POST_ORDER_SUCCESS, POST_ORDER_REQUEST, POST_ORDER_ERROR } from "../../utils/constants"
+import { ADD_BUN, ADD_INGREDIENT, MOVE_INGREDIENT, REMOVE_ITEM, ADD_TO_INGREDIENTS_MAP, ADD_TO_BUN_MAP, POST_ORDER_SUCCESS, POST_ORDER_REQUEST, POST_ORDER_ERROR } from "../../utils/constants"
 import { addBun, addIngredient, addToIngredientsMap, addToBunMap, postOrderSuccess, postOrderRequest, postOrderError } from '../actions/constructor-actions'
 
 
@@ -73,37 +73,3 @@ export const constructorReducer = (state = initialState, action) => {
       return state
   }
 }
-
-export const addToConstructor = item => dispatch => {
-  if (item.type === 'bun') {
-    dispatch(addToBunMap(item._id))
-    dispatch(addBun(item))
-  } else {
-    dispatch(addToIngredientsMap(item._id))
-    dispatch(addIngredient(item))
-  }
-}
-
-export const postOrder = data => dispatch => {
-  dispatch(postOrderRequest())
-  fetch(`${BASE_URL}/orders`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      'ingredients': data
-    }),
-  })
-    .then(res => res.json())
-    .then(res => {
-      if (res && res.success) {
-        dispatch(postOrderSuccess(res.order.number));
-      }
-    })
-    .catch(err => {
-      dispatch(postOrderError())
-      console.log(err)
-    })
-}
-
